Added the possibility for the keyboard to be debugged as well. I decided to have the keyboard to pass all commands to the debugger as well (just like with instructions). This allows debugger commands to be processed automatically, which is needed to fix the code to enable use of the teleporter.
To ensure a clean interface, created a DebuggerSupport interface containing the setDebugger/resetDebugger methods and the DEFAULT_DEBUGGER instance, and implemented it on Device and Keyboard.

Extended the DeviceDebugger interface with a debug(String) method to intercept commands. Changed both debug() methods to return a (possibly updated) value of their parameter (changed instruction or command string).

Finally updated Keyboard to ignore commands starting with '#' to allow for comment lines in the input file.